:memo: Recommend kernel version 2.6.32
Reasoning: the pre-compiled binaries that are downloaded by JuNest
during installation as well as binaries that are available from the Arch
repositories are compiled for Linux kernel 2.6.32. Therefore, with
kernel version 2.6.32 on the host OS no warnings, errors, or unexpected
crashes due to kernel version incompatibilities should be expected.

The recommended kernel version is now mentioned in the "Dependencies"
section of the README. In the "Troubleshooting" section about "Kernel
too old", more specific information about the problem and the workaround
using PRoot -k option is added.

Resolves: #172