:memo: Add Q&A entry about private futex warning
Since this problem can be pin-pointed to kernel versions that are older
than 2.6.22, a Q&A entry separate from the more general "Kernel too
old" problem may be justified.

This Q&A entry discusses the origin of the problem, some of the
symptoms (no symptoms for "simple" programs like vim, but warnings and
crashes for others), and concludes with a remark about possible (but
complicated) workarounds.